# secp256k1: Reduce EC operation normalizes.